# specs/contracts/group-class-booking.openapi.yaml
openapi: 3.0.3
info:
  title: Yoga Sys Group Class Booking API
  version: 0.1.0
  description: Contract for class catalog, weekly scheduling, booking write-off, and member account binding.
servers:
  - url: https://api.yoga-sys.local
security:
  - bearerAuth: []
...
paths:
  /courses:
    get:
      tags: [Courses]
      summary: List courses
      responses:
        '200':
          description: Course list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseListResponse' }
    post:
      tags: [Courses]
      summary: Create course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCourseRequest' }
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
  /courses/{courseId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Courses]
      summary: Get course
      responses:
        '200':
          description: Course returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
    patch:
      tags: [Courses]
      summary: Update or disable course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCourseRequest' }
      responses:
        '200':
          description: Course updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
  /rooms:
    get:
      tags: [Rooms]
      summary: List rooms
      responses:
        '200':
          description: Room list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomListResponse' }
    post:
      tags: [Rooms]
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoomRequest' }
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
  /rooms/{roomId}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      tags: [Rooms]
      summary: Get room
      responses:
        '200':
          description: Room returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
    patch:
      tags: [Rooms]
      summary: Update or disable room
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoomRequest' }
      responses:
        '200':
          description: Room updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Room' }
  /coaches:
    get:
      tags: [Coaches]
      summary: List coach profiles
      responses:
        '200':
          description: Coach list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachListResponse' }
    post:
      tags: [Coaches]
      summary: Create coach profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCoachRequest' }
      responses:
        '201':
          description: Coach created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Coach' }
  /coaches/{coachId}:
    parameters:
      - $ref: '#/components/parameters/CoachId'
    get:
      tags: [Coaches]
      summary: Get coach profile
      responses:
        '200':
          description: Coach returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Coach' }
    patch:
      tags: [Coaches]
      summary: Update or disable coach profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCoachRequest' }
      responses:
        '200':
          description: Coach updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Coach' }
...
  /members/{memberId}/account:
    post:
      tags: [MemberAccounts]
      summary: Create member password account
      parameters:
        - $ref: '#/components/parameters/MemberId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountBindingRequest' }
      responses:
        '201':
          description: Member account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountBinding' }
        '409': { description: Username or member binding already exists }
  /coaches/{coachId}/account:
    post:
      tags: [MemberAccounts]
      summary: Create coach password account
      parameters:
        - $ref: '#/components/parameters/CoachId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountBindingRequest' }
      responses:
        '201':
          description: Coach account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountBinding' }
        '409': { description: Username or coach binding already exists }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CoachId:
      name: coachId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MemberId:
      name: memberId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  schemas:
    CourseDifficulty:
      type: string
      enum: [all_levels, beginner, intermediate, advanced]
...
    CreateCourseRequest:
      type: object
      required: [name, durationMinutes]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        durationMinutes: { type: integer, minimum: 1, maximum: 480 }
        difficulty: { $ref: '#/components/schemas/CourseDifficulty' }
        description: { type: string, maxLength: 4000, nullable: true }
        coverUrl: { type: string, format: uri, nullable: true }
        enabled: { type: boolean, default: true }
    UpdateCourseRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        durationMinutes: { type: integer, minimum: 1, maximum: 480 }
        difficulty: { $ref: '#/components/schemas/CourseDifficulty' }
        description: { type: string, maxLength: 4000, nullable: true }
        coverUrl: { type: string, format: uri, nullable: true }
        enabled: { type: boolean }
    Course:
      type: object
      required: [id, name, durationMinutes, difficulty, enabled, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        durationMinutes: { type: integer }
        difficulty: { $ref: '#/components/schemas/CourseDifficulty' }
        description: { type: string, nullable: true }
        coverUrl: { type: string, format: uri, nullable: true }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CourseListResponse:
      type: object
      required: [items, total, skip, limit]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Course' } }
        total: { type: integer }
        skip: { type: integer }
        limit: { type: integer }
    CreateRoomRequest:
      type: object
      required: [name, capacity]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        capacity: { type: integer, minimum: 1, maximum: 500 }
        enabled: { type: boolean, default: true }
    UpdateRoomRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        capacity: { type: integer, minimum: 1, maximum: 500 }
        enabled: { type: boolean }
    Room:
      type: object
      required: [id, name, capacity, enabled, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        capacity: { type: integer }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RoomListResponse:
      type: object
      required: [items, total, skip, limit]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Room' } }
        total: { type: integer }
        skip: { type: integer }
        limit: { type: integer }
    CreateCoachRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 50 }
        avatarUrl: { type: string, format: uri, nullable: true }
        bio: { type: string, maxLength: 4000, nullable: true }
        specialtyCourseIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        enabled: { type: boolean, default: true }
    UpdateCoachRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 50 }
        avatarUrl: { type: string, format: uri, nullable: true }
        bio: { type: string, maxLength: 4000, nullable: true }
        specialtyCourseIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        enabled: { type: boolean }
    Coach:
      type: object
      required: [id, name, enabled, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        avatarUrl: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
        specialtyCourseIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CoachListResponse:
      type: object
      required: [items, total, skip, limit]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Coach' } }
        total: { type: integer }
        skip: { type: integer }
        limit: { type: integer }
...
    CreateAccountBindingRequest:
      type: object
      required: [username, initialPassword]
      properties:
        username: { type: string, minLength: 3, maxLength: 64, pattern: '^[a-z0-9_.-]+$' }
        initialPassword: { type: string, minLength: 8, maxLength: 128, format: password, writeOnly: true }
    AccountBinding:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { type: string, enum: [member, coach] }
        memberId: { type: string, format: uuid, nullable: true }
        coachProfileId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }